Fix handling of ZeroBits argument in NtAllocateVirtualMemory and friends
This patch makes NtAllocateVirtualMemory respect the ZeroBits argument,
whose functionality is derived from the following documentation for
NtAllocateVirtualMemory:
> The number of high-order address bits that must be zero in the base address
> of the section view. Used only when the operating system determines where
> to allocate the region, as when BaseAddress is NULL. Note that when
> ZeroBits is larger than 32, it becomes a bitmask.

and NtMapViewOfSection:
> Specifies the number of high-order address bits that must be zero in the
> base address of the section view. The value of this parameter must be less
> than 21 and is used only if BaseAddress is NULL—in other words, when the
> caller allows the system to determine where to allocate the view.

and from the documentation for LuaJIT v2.1's allocator in lj_alloc.c:
> /* Number of top bits of the lower 32 bits of an address that must be zero.
> ** Apparently 0 gives us full 64 bit addresses and 1 gives us the lower 2GB.
> */
> #define NTAVM_ZEROBITS 1

Thus the interpretation done here is:
If zero_bits is 0, use the full address space.
If zero_bits conflicts with the set alignment, such that the only possible
address is null, err.
Otherwise only the lower `32-zero_bits` bits of the address should be used.
The case where it functions as a bitmask is entirely ignored in this patch intentionally.

In addition, to implement this odd Windows API, Wine somehow has to
mmap an address below the maximum, which is not possible efficiently
on POSIX systems compared to the Windows API. Linux has the MAP_32BIT
flag, but it restricts the address space to just 1 GiB which is not
enough. For that reason, the implementation of this API in Wine
uses pseudorandom heuristics ripped from the LuaJIT project (which
also seems to be the only project to use this Windows API) to achieve
this. It is not optimal but it works. A proper implementation would
require extensions to the mmap API in the kernel.

This fixes all programs, that depend on 64-bit LuaJIT without GC64, not working,
e.g. the game Natural Selection 2, and programs that still use LuaJIT v2.0.

Signed-off-by: Las <las@protonmail.ch>